+++ phpmyfaq/src/phpMyFAQ/Configuration.php
@@ -19,6 +19,9 @@
 namespace phpMyFAQ;
 
 use Elasticsearch\Client;
+use Monolog\Handler\BrowserConsoleHandler;
+use Monolog\Handler\StreamHandler;
+use Monolog\Logger;
 use phpMyFAQ\Database\DatabaseDriver;
 
 /**
@@ -28,14 +31,13 @@ use phpMyFAQ\Database\DatabaseDriver;
  */
 class Configuration
 {
-    /**
-     * @var array
-     */
+    /** @var array */
     public array $config = [];
 
-    /**
-     * @var string
-     */
+    /** @var Logger */
+    public Logger $logger;
+
+    /** @var string */
     protected string $tableName = 'faqconfig';
 
     /**
@@ -46,6 +48,7 @@ class Configuration
     public function __construct(DatabaseDriver $database)
     {
         $this->setDb($database);
+        $this->setLogger();
     }
 
     /**
@@ -56,6 +59,19 @@ class Configuration
     public function setDb(DatabaseDriver $database): void
     {
         $this->config['core.database'] = $database;
+    }
+
+    /**
+     * Sets the Monolog logger instance, logs into a normal logfile
+     * If DEBUG is true, it logs to the browser console as well.
+     */
+    public function setLogger(): void
+    {
+        $this->logger = new Logger('phpmyfaq');
+        $this->logger->pushHandler(new StreamHandler(PMF_LOG_DIR, DEBUG ? Logger::DEBUG : Logger::WARNING));
+        if (DEBUG) {
+            $this->logger->pushHandler(new BrowserConsoleHandler());
+        }
     }
 
     /**