+++ tests/bootstrap.php
@@ -28,9 +28,11 @@ error_reporting(E_ALL | E_STRICT);
 //
 define('PMF_ROOT_DIR', dirname(__DIR__) . '/phpmyfaq');
 define('PMF_CONFIG_DIR', dirname(__DIR__) . '/phpmyfaq/config');
-define('PMF_TEST_DIR', __DIR__);
-define('IS_VALID_PHPMYFAQ', true);
-define('DEBUG', true);
+
+const PMF_LOG_DIR = __DIR__ . '/logs';
+const PMF_TEST_DIR = __DIR__;
+const IS_VALID_PHPMYFAQ = true;
+const DEBUG = true;
 
 $_SERVER['HTTP_HOST'] = 'https://localhost/';